Fix implicit narrowing conversion in compound assignment
https://github.com/arianne/stendhal/security/code-scanning/17
https://github.com/arianne/stendhal/security/code-scanning/18
https://github.com/arianne/stendhal/security/code-scanning/19
https://github.com/arianne/stendhal/security/code-scanning/20
https://github.com/arianne/stendhal/security/code-scanning/21
https://github.com/arianne/stendhal/security/code-scanning/22

+++ src/games/stendhal/client/sound/system/processors/VolumeAdjustor.java
@@ -1,6 +1,6 @@
 /* $Id$ */
 /***************************************************************************
- *                   (C) Copyright 2003-2010 - Stendhal                    *
+ *                   (C) Copyright 2003-2023 - Stendhal                    *
  ***************************************************************************
  ***************************************************************************
  *                                                                         *
@@ -133,7 +133,7 @@ public class VolumeAdjustor extends SignalProcessor
                 double vol   = volume + (volumeSegment * i / numSamples);
 
                 for(int c=0; c<channels; ++c) {
-					data[index + c] *= vol;
+					data[index + c] *= (float) vol;
 				}
             }
